moved getUsersSelection function and test to a service class
